+++ src/components/common/Pagination.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useRecoilState } from 'recoil';
-import { pageState } from '@src/types/pageState';
+import { pageState } from '@src/states/pageState';
 import Button from './PaginationButton';
 
 type PaginationProps = {
